For fluidsynth-2.1.8-win10-x64, needed to reorder shutdown sequence.
Otherwise a crash during cleanup would occur. Conceptually this makes
sense since we pass a the precontructed settings in to the synth
constructor. If we delete those settings prior to deleting the synth
it's evident that this triggers the exception.  This bug appears in
a debug-build of chuck+chugins and may be missed in normal (non-checked)
operation.