Rearchitect for proper threading. Bugfix on audioProcesser ownership.
The VST3 spec requires that IComponent::setupProcessing() is called
on the message thread. If you call it from a different thread, some
plugins may break.

Chugin methods are invoked in audiothread.

Host creates a "workerthread" into which the plugins are loaded
asynchronously.

Host delegates audiothread messages to worker thread with the
exception of tick. (WIP).